update: mv to pkg
